[refactor] 댓글에 jwt 토큰 방식 적용

+++ src/main/java/muit/backend/service/CommentServiceImpl.java
@@ -3,6 +3,8 @@ package muit.backend.service;
 import lombok.RequiredArgsConstructor;
 import muit.backend.apiPayLoad.code.status.ErrorStatus;
 import muit.backend.apiPayLoad.exception.GeneralException;
+import muit.backend.apiPayLoad.code.status.ErrorStatus;
+import muit.backend.apiPayLoad.exception.GeneralException;
 import muit.backend.converter.CommentConverter;
 import muit.backend.domain.entity.member.Comment;
 import muit.backend.domain.entity.member.Member;
@@ -42,9 +44,8 @@ public class CommentServiceImpl implements CommentService {
     //특정 게시물에 댓글 생성
     @Override
     @Transactional
-    public CommentReplyResponseDTO.CommentResponseDTO writeComment(CommentReplyRequestDTO.CommentRequestDTO requestDTO, Long postId) {
+    public CommentReplyResponseDTO.CommentResponseDTO writeComment(CommentReplyRequestDTO.CommentRequestDTO requestDTO, Long postId,Member member) {
         Post post = postRepository.findById(postId).orElseThrow(()->new GeneralException(ErrorStatus.POST_NOT_FOUND));
-        Member member = memberRepository.findById(requestDTO.getMemberId()).orElseThrow(()->new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
         Comment comment = CommentConverter.toComment(requestDTO, post, member);
 
         commentRepository.save(comment);
@@ -57,9 +58,8 @@ public class CommentServiceImpl implements CommentService {
     //특정 댓글에 대댓글 생성
     @Override
     @Transactional
-    public CommentReplyResponseDTO.ReplyResponseDTO writeReply(CommentReplyRequestDTO.ReplyRequestDTO requestDTO, Long commentId) {
+    public CommentReplyResponseDTO.ReplyResponseDTO writeReply(CommentReplyRequestDTO.ReplyRequestDTO requestDTO, Long commentId, Member member) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(()->new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
-        Member member = memberRepository.findById(requestDTO.getMemberId()).orElseThrow(()->new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
 
         Reply reply = CommentConverter.toReply(requestDTO, comment, member);
 
@@ -71,9 +71,15 @@ public class CommentServiceImpl implements CommentService {
     //특정 댓글 삭제
     @Override
     @Transactional
-    public CommentReplyResponseDTO.DeleteResultDTO deleteComment(String commentType, Long commentId) {
+    public CommentReplyResponseDTO.DeleteResultDTO deleteComment(String commentType, Long commentId, Member member) {
+
+
         if(commentType.equals("COMMENT")){
             Comment comment = commentRepository.findById(commentId).orElseThrow(()->new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
+            //작성자와 동일인인지 검사
+            if(comment.getMember()!=member){
+                throw(new GeneralException(ErrorStatus._FORBIDDEN));
+            }
             if(!comment.getReplyList().isEmpty()){
                 comment.deleteContent("삭제된 댓글입니다.");
             }else{
@@ -81,7 +87,11 @@ public class CommentServiceImpl implements CommentService {
             }
             comment.getPost().changeCommentCount(false);
         }else if(commentType.equals("REPLY")){
-            Reply reply = replyRepository.findById(commentId).orElseThrow(()->new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));
+            Reply reply = replyRepository.findById(commentId).orElseThrow(()->new GeneralException(ErrorStatus.COMMENT_NOT_FOUND));            //작성자와 동일인인지 검사
+            if(reply.getMember()!=member){
+                throw(new GeneralException(ErrorStatus._FORBIDDEN));
+            }
+
             replyRepository.deleteById(commentId);
             reply.getComment().getPost().changeCommentCount(false);
